implementing payout schedule first part, components and pages
