+++ frontend/src/components/areas/private/features/payout-settings/pages/payout-settings-bank-account-payout-schedule-page.tsx
@@ -1,0 +1,13 @@
+import PayoutScheduleForm from 'design-library/organisms/forms/payout-forms/payout-schedule-form/payout-schedule-form';
+import React from 'react';
+
+const PayoutSettingsBankAccountPayoutSchedulePage = () => {
+  return (
+    <div>
+      <h1>Payout Schedule</h1>
+      <PayoutScheduleForm />
+    </div>
+  );
+};
+
+export default PayoutSettingsBankAccountPayoutSchedulePage;
