+++ frontend/src/components/design-library/molecules/drawers/offer-drawer/components/offer-drawer-create.tsx
@@ -3,9 +3,9 @@ import { FormattedMessage } from 'react-intl';
 import Introduction from '../../../content/introduction/introduction';
 import IssueCard from '../../../cards/issue-card/issue-card';
 import SimpleInfo from '../../../../atoms/alerts/simple-info/simple-info';
-import DeliveryDate from '../../../../organisms/forms/delivery-date/delivery-date';
+import DeliveryDate from '../../../../organisms/forms/date-forms/delivery-date/delivery-date';
 import PickupTagList from '../../../lists/pickup-tag-list/pickup-tag-list';
-import PricePlan from '../../../../organisms/forms/price-plan-form/price-plan';
+import PricePlan from '../../../../organisms/forms/price-forms/price-plan-form/price-plan';
 import InputComment from '../../../../atoms/inputs/input-comment/input-comment';
 import OfferDrawerCheckboxes from './offer/offer-drawer-checkboxes';
 import InviteInput from './invite/invite-input';
