+++ frontend/src/components/design-library/organisms/forms/signup-forms/signup-signin/signup-signin.stories.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import SignupSignin from './signup-signin';
+
+export default {
+  title: 'Design Library/Organisms/Forms/SignupForms/SignupSignin',
+  component: SignupSignin
+};
+
+const Template = (args) => <SignupSignin {...args} />;
+
+export const Default = Template.bind({});
+Default.args = {};
