--- frontend/src/components/design-library/organisms/forms/signup-signin/signup-signin.stories.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import SignupSignin from './signup-signin';
-
-export default {
-  title: 'Design Library/Organisms/Forms/SignupSignin',
-  component: SignupSignin
-};
-
-const Template = (args) => <SignupSignin {...args} />;
-
-export const Default = Template.bind({});
-Default.args = {};
