+++ frontend/src/components/design-library/organisms/forms/subscribe-forms/subscribe-form/subscribe-form.stories.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import SubscribeForm from './subscribe-form';
+
+export default {
+  title: 'Design Library/Organisms/Forms/SubscribeForms/SubscribeForm',
+  component: SubscribeForm
+};
+
+const Template = (args) => <SubscribeForm {...args} />;
+
+export const Default = Template.bind({});
+Default.args = {
+  // Add default props here if any
+  render: true,
+  type: 'subscribe-form'
+};
