+++ frontend/src/components/design-library/organisms/layouts/bottom-bar/bottom.tsx
@@ -10,7 +10,7 @@ import {
 } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
 
-import SubscribeForm from '../../forms/subscribe-form/subscribe-form'
+import SubscribeForm from '../../forms/subscribe-forms/subscribe-form/subscribe-form'
 import StatsBar from '../../../molecules/sections/stats-bar/stats-bar'
 import SlackCard from './SlackCard'
 import GithubCard from './GithubCard'
