+++ frontend/src/components/design-library/pages/private/settings-pages/payout-settings-paypal/payout-settings-paypal.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import PaypalInfoForm from '../../../../organisms/forms/paypal-info-form/paypal-info-form';
+import PaypalInfoForm from '../../../../organisms/forms/payout-forms/paypal-info-form/paypal-info-form';
 import ProfileHeader from '../../../../molecules/headers/profile-main-header/profile-main-header';
 import { FormattedMessage } from 'react-intl';
